Spiketime Dependent Combining
Needs proper formatting in Line 15 referencing each outfile but it should look to see if each outfile has any spiketimes (i.e. was present at this day and firing and the corresponding session the directory was generated) and only add outfiles that have spiketimes (i.e. sessions the cell was present and firing)